src/pyproject.toml
[project]
name = "financas-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4",
    "django>=5.0,<6.0",
    "django-cors-headers>=4.4",
    "django-environ>=0.11",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "psycopg[binary]>=3.2",
    "redis>=5.0",
    "openai>=1.50",
]
